* App.xaml.cs: fix problem with posting to piecevente and   pieceventeligne
* Strings.cs:
* RestServices.cs:
* IRestServices.cs:
* SellDetailsPageModel.cs:

+++ PFE/PFE/PageModels/SellDetailsPageModel.cs
@@ -35,11 +35,15 @@ namespace PFE.PageModels
         {
 
 
+            if(_restservices.PostSellLignes(productList)){
 
-            if(_dataService.RemoveSellElements()){
-                productList.Clear();
-                _dialogService.ShowMessage("working on valiid", true);
+                if (_dataService.RemoveSellElements())
+                {
+                    productList.Clear();
+                    _dialogService.ShowMessage("Done", false);
+                }
             }
+
         }
 
         private void _delete(object obj)
